Update on "[FX] Shape propagation example"
The example in the internal repo bitrotted, so here's an updated version with test coverage

Differential Revision: [D24024606](https://our.internmc.facebook.com/intern/diff/D24024606)

[ghstack-poisoned]
