+++ test/test_fx.py
@@ -7,6 +7,7 @@ import copy
 from pathlib import Path
 from torch.fx import symbolic_trace, Proxy, Node, GraphModule, Tracer, Graph
 from torch.fx.experimental import GraphManipulation
+from torch.fx.experimental import shape_prop
 
 from torch.fx.proxy import TraceError
 
@@ -733,7 +734,6 @@ class TestFX(JitTestCase):
         with self.assertRaisesRegex(RuntimeError, 'was used before it has been defined'):
             graph.lint()
 
-    @skipIfNoTorchVision
     def test_example_shape_prop(self):
         class TestCase(torch.nn.Module):
             def __init__(self):
@@ -745,51 +745,6 @@ class TestFX(JitTestCase):
                 return torch.neg(self.submod(x.relu() + self.attr))
         tc = TestCase()
         tc_traced = symbolic_trace(tc)
-
-        class ShapeProp:
-            def __init__(self, mod):
-                self.mod = mod
-                self.graph = mod.graph
-                self.modules = dict(self.mod.named_modules())
-
-            def propagate(self, *args):
-                args_iter = iter(args)
-                env : Dict[str, Node] = {}
-
-                def load_arg(a):
-                    return torch.fx.graph.map_arg(a, lambda n: env[n.name])
-
-                def fetch_attr(target : str):
-                    target_atoms = target.split('.')
-                    attr_itr = self.mod
-                    for i, atom in enumerate(target_atoms):
-                        if not hasattr(attr_itr, atom):
-                            raise RuntimeError(f"Node referenced nonexistant target {'.'.join(target_atoms[:i])}")
-                        attr_itr = getattr(attr_itr, atom)
-                    return attr_itr
-
-                for node in self.graph.nodes:
-                    if node.op == 'placeholder':
-                        result = next(args_iter)
-                    elif node.op == 'get_attr':
-                        result = fetch_attr(node.target)
-                    elif node.op == 'call_function':
-                        result = node.target(*load_arg(node.args), **load_arg(node.kwargs))
-                    elif node.op == 'call_method':
-                        self_obj, *args = load_arg(node.args)
-                        kwargs = load_arg(node.kwargs)
-                        result = getattr(self_obj, node.target)(*args, **kwargs)
-                    elif node.op == 'call_module':
-                        result = self.modules[node.target](*load_arg(node.args), **load_arg(node.kwargs))
-
-                    if isinstance(result, torch.Tensor):
-                        node.shape = result.shape
-                        node.dtype = result.dtype
-
-                    env[node.name] = result
-
-                return load_arg(self.graph.result)
-
         ref_out = tc_traced(torch.rand(3, 4))
 
         # Make sure we're testing all opcodes
@@ -799,7 +754,7 @@ class TestFX(JitTestCase):
         self.assertEqual(opcodes, set(['placeholder', 'get_attr', 'call_function', 'call_method', 'call_module']))
 
         # Test shape propogation and make sure results match actual
-        ShapeProp(tc_traced).propagate(torch.rand(3, 4))
+        shape_prop.ShapeProp(tc_traced).propagate(torch.rand(3, 4))
         self.assertEqual(tc_traced.graph.result.shape, ref_out.shape)
